#729 Renamed SpaceRaceClass to SpaceRace and SpaceRace to SpaceRaceEnum.
Work still in progress.
