+++ src/main/java/org/openRealmOfStars/player/race/SpaceRaceType.java
@@ -1,0 +1,46 @@
+package org.openRealmOfStars.player.race;
+/*
+ * Open Realm of Stars game project
+ * Copyright (C) 2024 Tuomo Untinen
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, see http://www.gnu.org/licenses/
+ */
+
+/** Space race type */
+public enum SpaceRaceType {
+  /** Regular playable space race */
+  REGULAR,
+  /** Special non playable space race, but has diplomacy */
+  SPACE_PIRATE,
+  /** Special non playable space race, but no diplomacy */
+  SPACE_MONSTER;
+
+  /**
+   * Get space race type by string.
+   * @param value String value
+   * @return Space race type.
+   */
+  public static SpaceRaceType getByString(final String value) {
+    if (value.equalsIgnoreCase("regular")) {
+      return REGULAR;
+    }
+    if (value.equalsIgnoreCase("space_pirate")) {
+      return SPACE_PIRATE;
+    }
+    if (value.equalsIgnoreCase("space_monster")) {
+      return SPACE_MONSTER;
+    }
+    throw new IllegalArgumentException("Unexpected space race type: " + value);
+  }
+}
